+++ src/main/java/pl/edu/pw/mini/ingreedio/api/user/dto/UserDto.java
@@ -5,6 +5,6 @@ import lombok.Builder;
 import pl.edu.pw.mini.ingreedio.api.ingredient.dto.IngredientDto;
 
 @Builder
-public record UserDto(Long userId, String email, String displayName,
+public record UserDto(int userId, String email, String displayName,
                       Set<IngredientDto> likedIngredients, Set<IngredientDto> allergens) {
 }
